fix: duplicate denotes for same org-roam file
org-roam does this bullshit where any org-subtree in an org-roam note file is
treated as a separate org-roam node. This make migrated denotes create duplicate
notes because we operate directly on org-roam files.

This commit fixes it by operating on org-roam files instead of asking org-roam
nicely to provide a list of org-roam nodes.