nm/denote-to-logseq: Wrote RX matching denote-link
I am still trying to wrap my head around opening and writing buffers.  For starters,
I wrote a RX expression that I believe to be able to 100% identify a denote link.

This is important because the DENOTE and LOGSEQ link format seems to be quite
different. DENOTE has [[denote:ID][DESC]], LOGSEQ has [[FILE TITLE]] without
description.  Thus I think it is important to be specific with RX.

Next up is the identification of the file DENOTE links TO, getting this file's TITLE.