nm/denote-to-logseq: Getting closer I guess?
So I have written NM--CONVERT-DENOTE-LINKS-TO-LOGSEQ that does most of the things I
want corrctly, whereby I mean that all values in the LET-expressions are correct.
However, there is an issue with matching and replacing that I don't quite understand.

Currently, I have one magic variable baked in, MY-DENOTE-DIRECTORY.  I will have to
find a way how to eliminate it, though I think this is less important than the
matching/replacing part.

Also, the file is a total mess overall, will have to clean up soon™.