nm/denote-to-logseq: Approaching a working prototype
So I found out that I can use denote's internal machinery to great effect, which I
did.  Now, I believe to have all the pieces I need in order to replace a denote link
by a logseq link.

However, for a reason that I don't understand, the combination of `re-search-forward`
and `replace-match` does not work as I would expect it to work.

I will have to investigate this further tomorrow.