nm/denote-to-logseq: Trying to resolve the match/replace issue
So I thought I might find an ORG-function with which I can access the link under
point and just change the link contents, without resorting to regex replacement.
Unfortunately, I was unable to find that.

I honestly don't understand the error that's thrown (args out of range) and how to
mitigate it.  Strangely enough, I am able to successfully replace ONE link when I
insert a `re-search-backward` before the `replace-match`.  However, when I wrap the
whole shebang in a `while` instead of an `if`, the execution does not terminate.