nm/denote-to-logseq: fix: Proper new link inserting and and directory substitution
Testing your new code, no files would be written to a new directory.  However, the
links would be changed in the old directory (thank god I have used "test-*"
directories!), albeit without inserting a space after.

I thus checked the functioning of `nm--migrate-denote-files-to-logseq` and found that
`expand-file-name` does not behave as expected in combination with your rendition of
`string-replace`.  Using LET to input values for `denote-directory` and
`logseq-directory`, I found that DEST evaluates to ./path-to-file.org.

I then remembered that there is a native function that returns just the file name
without directory, `file-name-nondirectory`.  Using these changes, I was able to
successfully migrate from TEST-DENOTE to TEST-LOGSEQ :)

I don't know how exactly this works, but I know that it has not worked before and
works afterwards with the same input values.

Going forward, what do you think is required to improve on the version that is
already present?  One thing I have in mind but with which I don't have any experience
whatsoever is emitting error messages when stuff doesn't work as expected.  Then, I
feel like returning a list containing N NIL where N is the number of files is kind of
noisy.

Furthermore, I believe the next step would be to check if there is a way to refactor
so that there is less redundance for the two migration functions (roam -> denote,
denote -> logseq).  Again, that is an idea, but I am not experienced with this kind
of stuff.

Two closing thoughts: first off, at what point do you think it would make sense to
post and update to reddit (could increase exposure and more skilled people joining
in)? Second, what would be required to make the program, as is, usable as a package?
I feel like the minimum requirement for that would be the capacity to define
LOGSEQ-DIRECTORY and DENOTE-DIRECTORY and to emit an error message when they are not
defined or no files are found there.