nm/denote-to-logseq: Cleaner code
I removed CONTENTS in LET* because it wasn't used and served no purpose.

I implemented the solution proposed by yantar92, which meant to replace
`delete-region` and `insert` with `replace-region-contents`.

Finally, I thought it a good idea to also steal (from yantar92) the
`org-fill-paragraph` function at the end so that the file doesn't look ugly after
replacing a long (denote) link with a much shorter (logseq) link.  Dunno if that
matters though.