feature: Detect a silo'd file and act accordingly (not finished)
So I wrote functions that detect if file is not in DENOTE-DIRECTORY root but in a
subdirectory, and act accordingly: produce a fitting LOGSEQ-SILO (LOGSEQ-DIRECTORY +
silo/subdir) string, mkdir (if not existing), write file to DEST.

After a short moment of triumph I realized that even though the logic of this
function is sound, it does not work well with the logic in place for
NM--MIGRATE-DENOTE-TO-LOGSEQ.  This function relies on the fact that the
DENOTE-DIRECTORY contains org files, which it doesn't, both in my use case and the
present test case.

This cannot be too complicated, but I don't have the resources to finish it today.