docs: Known limitations in denote-to-logseq migrator
There are quite a number of limitations in the current iteration of the
denote-to-logseq migrator that should be visible to users.  Not entirely certain if
this is the proper way to convey things in a readme, but I feel I could not have been
more concise without leaving out vital stuff.